Update README Move string formatting to a utility struct
