+++ HasanPrestoQ/ManagersSpecialsViewController.swift
@@ -100,6 +100,4 @@ class ManagersSpecialsViewController: UICollectionViewController, UICollectionVi
         return headerView
     }
 
-
 }
-
